chore: Copy data to the code dir, add markdown comments to hopfield_Jupyter.ipynb, correct minor typos
To make the Python code for Hopfield Network run straight away when launched from the Hopfield_Network dir, the data had to be copied to the same directory, or the path in the code should be altered accordingly (and this should be in a different format for Windows and Unix-like systems). So I just copied the data to the code dir.